Share feature implementation
This commit should greatly cover our criteria for the "shared" feature that is required of our application.

I've introduced a new migration, and a new template called "share" I've extensively rerouted the routes to accomodate for the sharing changes.

The system I've introduced works as follows:

Users are assigned usernames of their choosing, usernames can be searched through a search feature in share.html if a valid username is inputted it offers their username and you can then choose to send a "friend request"

friend requests can be declined, but when accepted allow both parties to view each others data (theoretically as this data view function goes only so far as a button at the moment)

I've introduced a new model in models and also a new form (necessary for the search function)